Updates Cart and CartItems Bean and DAO: use of BigDecimal. CartBean: added setAttribute to put or replace items

+++ src/main/java/com/unisa/store/tsw_project/model/DAO/CartDAO.java
@@ -17,7 +17,7 @@ public class CartDAO {
             if (rs.next()) {
                 CartBean c = new CartBean();
                 c.setId_cart(rs.getInt(1));
-                c.setTotal(rs.getDouble(2));
+                c.setTotal(rs.getBigDecimal(2));
                 c.setId_client(rs.getInt(3));
                 c.setDiscount_code(rs.getString(4));
                 return c;
@@ -31,7 +31,7 @@ public class CartDAO {
             PreparedStatement ps = con.prepareStatement("INSERT INTO cart " +
                     "(total, id_client, discount_code) VALUES(?, ?, ?)",
                     PreparedStatement.RETURN_GENERATED_KEYS);
-            ps.setDouble(1, cart.getTotal());
+            ps.setBigDecimal(1, cart.getTotal());
             ps.setInt(2, cart.getId_client());
             ps.setString(3, cart.getDiscount_code());
             if (ps.executeUpdate() != 1) {
@@ -47,7 +47,7 @@ public class CartDAO {
         try (Connection con = ConPool.getConnection()) {
             PreparedStatement ps = con.prepareStatement("UPDATE cart " +
                     "SET total=?, id_client=?, discount_code=? WHERE id_cart=?");
-            ps.setDouble(1, cart.getTotal());
+            ps.setBigDecimal(1, cart.getTotal());
             ps.setInt(2, cart.getId_client());
             ps.setString(3, cart.getDiscount_code());
             ps.setInt(4, cart.getId_cart());
